1. Move my files from messagesConverter to converter 2. Move files from enums\* to type 3. Correct my branch MESS-11,6 Now application starts and you can work with your branches
