+++ src/main/java/com/itea/messenger/controller/MessagesController.java
@@ -16,30 +16,25 @@ public class MessagesController {
     private static final Logger log = LogManager.getLogger(MessagesController.class);
     private final MessagesService messagesService;
 
-//    @SneakyThrows
-
-    @PostMapping(value = "/save")
+    @PostMapping
     public MessagesDto saveMessages(@RequestBody MessagesDto messageDto) {
         log.info("Handling save message: " + messageDto);
         return messagesService.saveMessage(messageDto);
     }
 
-
-    @GetMapping("/getById")
+    @GetMapping("{Id}")
     public MessagesDto getMessageById(@RequestParam Long messageId) {
         log.info("Handling get message by ID: " + messageId);
         return messagesService.getMessageById(messageId);
     }
 
-
-    @GetMapping("/getAllByChatId")
+    @GetMapping("/chat")
     public List<MessagesDto> getAllMessagesByChatId(@RequestParam Long chatId) {
         log.info("Handling get all messages by chat ID: " + chatId);
         return messagesService.getAllMessagesByChatId(chatId);
     }
 
-
-    @DeleteMapping("/delete")
+    @DeleteMapping
     public void deleteMessage(@RequestParam Long messageId) {
         log.info("Handling delete message by ID: " + messageId);
         messagesService.deleteMessage(messageId);
