+++ src/main/java/com/itea/messenger/dto/MessagesDto.java
@@ -2,8 +2,8 @@ package com.itea.messenger.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import java.time.LocalDateTime;
 
+import java.time.LocalDateTime;
 
 @Data
 @Builder
@@ -13,8 +13,6 @@ public class MessagesDto {
     private Long userId;
     private String messageText;
     private Long fileId;
-
     private LocalDateTime dateTime;
 //    private Long statusId;
-
 }
