+++ src/main/java/com/itea/messenger/entity/StatusLinks.java
@@ -1,6 +1,6 @@
 package com.itea.messenger.entity;
 
-import com.itea.messenger.enums.messageStatus.MessageStatus;
+import com.itea.messenger.type.MessageStatus;
 import lombok.Data;
 import lombok.NoArgsConstructor;
