--- src/main/java/com/itea/messenger/enums/fileType/FileTypes.java
@@ -1,5 +1,0 @@
-package com.itea.messenger.enums.fileType;
-
-public enum FileTypes {
-    UNKNOWN, IMAGES, SOUND, WORD_DOCUMENT, PDF_DOCUMENT
-}
