--- src/main/java/com/itea/messenger/enums/messageStatus/MessageStatus.java
@@ -1,12 +1,0 @@
-package com.itea.messenger.enums.messageStatus;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum MessageStatus {
-    READ,
-    UNREAD,
-    SENT,
-    DELIVERED,
-    DELETED,
-    EDITED
-}
