--- src/main/java/com/itea/messenger/enums/messageStatus/PostgreSQLEnumMessageStatus.java
@@ -1,21 +1,0 @@
-package com.itea.messenger.enums.messageStatus;
-
-import org.hibernate.HibernateException;
-import org.hibernate.engine.spi.SharedSessionContractImplementor;
-import org.hibernate.type.EnumType;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.sql.Types;
-
-public class PostgreSQLEnumMessageStatus extends EnumType<MessageStatus> {
-    @Override
-    public void nullSafeSet(PreparedStatement ps, Object obj, int index,
-                            SharedSessionContractImplementor session) throws HibernateException, SQLException {
-        if (obj == null) {
-            ps.setNull(index, Types.OTHER);
-        } else {
-            ps.setObject(index, obj.toString(), Types.OTHER);
-        }
-    }
-}
