--- src/main/java/com/itea/messenger/enums/messageStatus/package-info.java
@@ -1,4 +1,0 @@
-@org.hibernate.annotations.TypeDef(name = "message_status", typeClass = PostgreSQLEnumMessageStatus.class)
-
-package com.itea.messenger.enums.messageStatus;
-
