+++ src/main/java/com/itea/messenger/type/FileTypes.java
@@ -1,0 +1,5 @@
+package com.itea.messenger.type;
+
+public enum FileTypes {
+    UNKNOWN, IMAGE, SOUND, WORD_DOCUMENT, PDF_DOCUMENT
+}
