+++ src/main/java/com/itea/messenger/type/MessageStatus.java
@@ -1,13 +1,10 @@
 package com.itea.messenger.type;
 
-public class messageStatus {
-
-    public enum MessageStatus {
+public enum MessageStatus {
         READ,
         UNREAD,
         SENT,
         DELIVERED,
         DELETED,
         EDITED
-    }
 }
