+++ src/main/java/com/itea/messenger/type/package-info.java
@@ -1,5 +1,4 @@
 @org.hibernate.annotations.TypeDef(name = "file_types_enum", typeClass = PostgreSQLEnumFileTypes.class)
 @org.hibernate.annotations.TypeDef(name = "message_status", typeClass = PostgreSQLEnumMessageStatus.class)
 
-package com.itea.messenger.enums.fileType;
-
+package com.itea.messenger.type;
